// src/app/model/common/AnagraficaGradautorie.ts
export class AnagraficaGraduatorie {
    idAnagraficaGraduatoria: number;
    codAnagraficaGraduatoria: string;
    codAnnoScolastico: string;
    codOrdineScuola: string;
    dataInizioIscrizioni: string;
    dataScadenzaIscrizioni: string;
    dataFineIscrizioni: string;
    dataScadenzaGraduatoria: string;
    dataScadenzaRicorsi: string;
}